refactor: move task pool into task module
