+++ src/global_state.rs
@@ -14,11 +14,7 @@ pub(crate) mod task;
 mod trace;
 mod workspace_state;
 
-use std::{
-    collections::HashMap,
-    panic::{self, AssertUnwindSafe},
-    time::Instant,
-};
+use std::time::Instant;
 
 use crossbeam_channel::{Receiver, Sender, unbounded};
 use hir::base_db::{
@@ -35,12 +31,7 @@ use parking_lot::{Mutex, RwLock};
 use project_model::Workspace;
 use rustc_hash::{FxHashMap, FxHashSet};
 use triomphe::Arc;
-use utils::{
-    cancellation::CancellationToken,
-    excl_task::ExclTask,
-    lines::LineEnding,
-    thread::{Pool, ThreadIntent},
-};
+use utils::{cancellation::CancellationToken, excl_task::ExclTask, lines::LineEnding};
 use vfs::{self, FileId, Vfs, notify::NotifyHandle};
 
 #[cfg(test)]
@@ -53,106 +44,11 @@ use self::{
     main_loop::DiagnosticPublishKey,
     mem_docs::MemDocs,
     snapshot::GlobalStateSnapshot,
-    task::Task,
+    task::{Task, TaskPool},
     trace::LspTrace,
     workspace_state::WorkspaceVfsReadiness,
 };
 use crate::config::{Config, ConfigError};
-
-pub(crate) struct TaskPool<T> {
-    pub(crate) sender: Sender<T>,
-    pub(crate) pool: Pool,
-    lifecycle_cancel: CancellationToken,
-    request_cancel_tokens: HashMap<lsp_server::RequestId, CancellationToken>,
-}
-
-impl<T> TaskPool<T> {
-    pub(crate) fn new_with_threads_num(sender: Sender<T>, threads_num: usize) -> TaskPool<T> {
-        TaskPool {
-            sender,
-            pool: Pool::new(threads_num),
-            lifecycle_cancel: CancellationToken::new(),
-            request_cancel_tokens: HashMap::new(),
-        }
-    }
-
-    pub(crate) fn task_token(&self) -> CancellationToken {
-        self.lifecycle_cancel.child_token()
-    }
-
-    pub(crate) fn register_request(
-        &mut self,
-        request_id: lsp_server::RequestId,
-    ) -> CancellationToken {
-        let token = self.task_token();
-        self.request_cancel_tokens.insert(request_id, token.clone());
-        token
-    }
-
-    pub(crate) fn request_token(
-        &self,
-        request_id: &lsp_server::RequestId,
-    ) -> Option<CancellationToken> {
-        self.request_cancel_tokens.get(request_id).cloned()
-    }
-
-    pub(crate) fn complete_request(&mut self, request_id: &lsp_server::RequestId) {
-        self.request_cancel_tokens.remove(request_id);
-    }
-
-    pub(crate) fn cancel_request(&mut self, request_id: &lsp_server::RequestId) {
-        if let Some(token) = self.request_cancel_tokens.remove(request_id) {
-            token.cancel();
-        }
-    }
-
-    pub(crate) fn cancel_all(&mut self) {
-        self.lifecycle_cancel.cancel();
-        self.request_cancel_tokens.clear();
-    }
-
-    pub(crate) fn spawn_and_send<F>(&mut self, intent: ThreadIntent, task: F)
-    where
-        F: FnOnce() -> T + Send + 'static,
-        T: Send + 'static,
-    {
-        self.pool.spawn(intent, {
-            let sender = self.sender.clone();
-            move || match panic::catch_unwind(AssertUnwindSafe(task)) {
-                Ok(task) => {
-                    if sender.send(task).is_err() {
-                        tracing::debug!("task result dropped because main loop receiver is closed");
-                    }
-                }
-                Err(panic) => log_task_panic(panic),
-            }
-        })
-    }
-
-    pub(crate) fn spawn_and_send_cps<F>(&mut self, intent: ThreadIntent, task: F)
-    where
-        F: FnOnce(Sender<T>) + Send + 'static,
-        T: Send + 'static,
-    {
-        self.pool.spawn(intent, {
-            let sender = self.sender.clone();
-            move || {
-                if let Err(panic) = panic::catch_unwind(AssertUnwindSafe(|| task(sender))) {
-                    log_task_panic(panic);
-                }
-            }
-        })
-    }
-}
-
-fn log_task_panic(panic: Box<dyn std::any::Any + Send>) {
-    let message = panic
-        .downcast_ref::<String>()
-        .map(String::as_str)
-        .or_else(|| panic.downcast_ref::<&str>().copied())
-        .unwrap_or("unknown panic payload");
-    tracing::error!(message, "background task panicked");
-}
 
 pub(crate) struct Handle<H, C> {
     pub(crate) handle: H,
@@ -331,34 +227,5 @@ impl GlobalState {
 
     pub(crate) fn cancel_all_tasks(&mut self) {
         self.task_pool.handle.cancel_all();
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::TaskPool;
-
-    #[test]
-    fn task_pool_request_cancel_signals_registered_token() {
-        let (sender, _receiver) = crossbeam_channel::unbounded::<()>();
-        let mut pool = TaskPool::new_with_threads_num(sender, 0);
-        let request_id = lsp_server::RequestId::from(7);
-        let token = pool.register_request(request_id.clone());
-
-        pool.cancel_request(&request_id);
-
-        assert!(token.is_cancelled());
-        assert!(pool.request_token(&request_id).is_none());
-    }
-
-    #[test]
-    fn task_pool_lifecycle_cancel_signals_child_tokens() {
-        let (sender, _receiver) = crossbeam_channel::unbounded::<()>();
-        let mut pool = TaskPool::new_with_threads_num(sender, 0);
-        let token = pool.task_token();
-
-        pool.cancel_all();
-
-        assert!(token.is_cancelled());
     }
 }
